# api/docs/swagger.yaml
definitions:
  controllers.request:
    properties:
      expiry:
        type: integer
      short:
        type: string
      url:
        type: string
    type: object
  controllers.response:
    properties:
      expiry:
        type: integer
      rate_limit:
        type: integer
      rate_limit_reset:
        type: integer
      short:
        type: string
      url:
        type: string
    type: object
info:
  contact:
    email: your@mail.com
    name: API Support
  description: This is an auto-generated API Docs.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: API
  version: "1.0"
paths:
  /api/v1/resolve/{url}:
    get:
      consumes:
      - '*/*'
      description: returns not shortened url.
      parameters:
      - description: shortened url
        in: path
        name: url
        required: true
        type: string
      responses:
        "301":
          description: redirected
          schema:
            type: string
        "400":
          description: desc
          schema:
            additionalProperties: true
            type: object
        "404":
          description: desc
          schema:
            additionalProperties: true
            type: object
        "500":
          description: desc
          schema:
            additionalProperties: true
            type: object
      summary: Returns original URL if exists.
      tags:
      - root
  /api/v1/shorten:
    post:
      consumes:
      - application/json
      description: returns not shortened url.
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.response'
        "403":
          description: desc
          schema:
            additionalProperties: true
            type: object
        "503":
          description: desc
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Returns original URL if exists.
      tags:
      - root
  /api/v1/token/new:
    get:
      consumes:
      - application/json
      description: Create a new access token.
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: create a new access token
      tags:
      - Token
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"